Add support price calculation and separate functions
