+++ static/js/src/advantage/subscribe/listeners/form-event-listeners.js
@@ -1,0 +1,53 @@
+import {
+  changeFeature,
+  changeQuantity,
+  changeSupport,
+  changeType,
+  changeVersion,
+  changeBilling,
+} from "../reducers/form-reducer";
+
+const inputs = [
+  {
+    action: changeType,
+    name: "type",
+  },
+  {
+    action: changeVersion,
+    name: "version",
+  },
+  {
+    action: changeFeature,
+    name: "feature",
+  },
+  {
+    action: changeSupport,
+    name: "support",
+  },
+  {
+    action: changeQuantity,
+    name: "quantity",
+  },
+];
+
+const form = document.querySelector(".js-shop-form");
+
+export default function initFormInputs(store) {
+  function initInputs(action, name) {
+    const inputs = form.querySelectorAll(`input[name='${name}']`);
+    inputs.forEach((input) => {
+      input.addEventListener("input", (e) => {
+        store.dispatch(action(e.target.value));
+      });
+    });
+  }
+
+  inputs.forEach((section) => {
+    initInputs(section.action, section.name);
+  });
+
+  const billingSelect = form.querySelector("#billing-period");
+  billingSelect.addEventListener("change", (e) => {
+    store.dispatch(changeBilling(e.target.value));
+  });
+}
