+++ static/js/src/advantage/subscribe/renderers/main-renderer.js
@@ -1,0 +1,5 @@
+import RenderForm from "./form-renderer";
+
+export default function render(state) {
+  RenderForm(state.form);
+}
