+++ static/js/src/advantage/subscribe/renderers/version-details.js
