+++ templates/advantage/subscribe/form/_tech_support.html
@@ -24,7 +24,7 @@
             loading="auto", 
             ) | safe
         }}
-        <p>$225 per machine per year</p>
+        <p id="essential-support-costs">No additional costs</p>
       </div>
 
       <div class="p-card--radio--support js-radio">
@@ -49,7 +49,7 @@
             loading="auto", 
             ) | safe 
           }}
-        <p>$750 per machine per year</p>
+        <p id="standard-support-costs">$750 per machine per year</p>
       </div>
 
       <div class="p-card--radio--support js-radio">
@@ -74,7 +74,7 @@
             loading="auto", 
             ) | safe 
           }}
-        <p>$1,500 per machine per year</p>
+        <p id="advanced-support-costs">$1,500 per machine per year</p>
       </div>
     </div>
     <div class="col-12">
